Doc site generation

+++ purge.ps1
@@ -1,12 +1,14 @@
 Write-Host 'Always wise to purge the rabble'
 
-$folderNames = 'obj', 'bin', 'CoverageResults', 'TestResult*'
-
+$folderNames = 'obj', 'bin', 'CoverageResults', 'TestResult', 'TestResult*'
 foreach ($folderName in $folderNames) {
     $folders = Get-ChildItem -Path $foldername -recurse
     foreach ($folder in $folders) {
-      Write-Host 'Deleting ' $folder.FullName;
-      Remove-Item -Path $folder.FullName -recurse -Force
+        if (Test-Path $folder.FullName)
+        {
+          Write-Host 'Deleting ' $folder.FullName;
+          Remove-Item -Path $folder.FullName -recurse -Force
+        }
     }
 }
 
@@ -15,8 +17,11 @@ foreach ($fileName in $fileNames) {
     $files = Get-ChildItem $fileName -recurse
     foreach ($file in $files)
     {
-      Write-Host 'Deleting ' file.FullName;
-      Remove-Item -Path $file.FullName -recurse -Force
+        if (Test-Path file.FullName)
+        {
+          Write-Host 'Deleting ' file.FullName;
+          Remove-Item -Path $file.FullName -recurse -Force
+        }
     }
 }
 
@@ -38,6 +43,16 @@ if (Test-Path *.binlog) {
 if (Test-Path TestResults) {
   Write-Host "Deleting TestResults"
   Remove-Item -Path TestResults -recurse -Force
+}
+
+if (Test-Path docs\api) {
+  Write-Host "Deleting docfx extracted api documentation"
+  Remove-Item -Path docs\api -recurse -Force
+}
+
+if (Test-Path docs\_site) {
+  Write-Host "Deleting docfx generated site"
+  Remove-Item -Path docs\_site -recurse -Force
 }
 
 Write-Host 'Done'
